extract quotes
